Restructure tests with fixtures and focused modules
Apply the test-improvement policy from the refactor plan:

- Add infrastructure::test_fixture with expression/value builders so
  Arrange sections read like the expression under test
  (add(int(1), mul(int(2), int(3))) instead of nested Box::new)
- Split monolithic test files into per-concern modules:
  evaluator 1786 lines -> 11 files, parser 1215 -> 11 files,
  type_checker 750 -> 6 files, CLI integration 1404 -> 13 files
- Delete 24 trivial tests that only exercised derived impls
  (Clone/PartialEq/Debug), including all of ast_test.rs
- Fix clippy approx_constant/identity_op findings in tests;
  cargo clippy --workspace --all-targets -- -D warnings is now clean

356 tests pass (380 before minus the 24 trivial deletions).
All assertions and expected values of remaining tests are unchanged.
