+++ ccc/domain/src/error_test.rs
@@ -101,65 +101,4 @@ mod tests {
             panic!("Expected TypeCheck variant");
         }
     }
-
-    #[test]
-    fn errors_are_distinct_variants() {
-        // Arrange
-        let scan = CccError::scan("x");
-        let parse = CccError::parse("x");
-        let eval = CccError::eval("x");
-        let type_check = CccError::type_check("x");
-
-        // Act & Assert
-        assert_ne!(scan, parse);
-        assert_ne!(parse, eval);
-        assert_ne!(scan, eval);
-        assert_ne!(type_check, scan);
-        assert_ne!(type_check, parse);
-        assert_ne!(type_check, eval);
-    }
-
-    #[test]
-    fn same_variant_same_message_are_equal() {
-        // Arrange
-        let a = CccError::scan("err");
-        let b = CccError::scan("err");
-
-        // Act & Assert
-        assert_eq!(a, b);
-    }
-
-    #[test]
-    fn same_variant_different_message_are_not_equal() {
-        // Arrange
-        let a = CccError::scan("err1");
-        let b = CccError::scan("err2");
-
-        // Act & Assert
-        assert_ne!(a, b);
-    }
-
-    #[test]
-    fn error_implements_std_error() {
-        // Arrange
-        let err: Box<dyn std::error::Error> = Box::new(CccError::eval("test"));
-
-        // Act
-        let result = err.to_string();
-
-        // Assert
-        assert!(result.contains("eval error"));
-    }
-
-    #[test]
-    fn debug_format_is_available() {
-        // Arrange
-        let err = CccError::parse("bad input");
-
-        // Act
-        let debug = format!("{err:?}");
-
-        // Assert
-        assert!(debug.contains("Parse"));
-    }
 }
