+++ ccc/infrastructure/src/lib.rs
@@ -1,3 +1,6 @@
 pub mod evaluator;
 pub mod parser;
 pub mod type_checker;
+
+#[cfg(test)]
+pub(crate) mod test_fixture;
